refactor(checkout-button): CHECKOUT-3011 Load checkout object only when required
