+++ src/checkout-buttons/create-checkout-button-initializer.ts
@@ -1,7 +1,4 @@
-import { createRequestSender } from '@bigcommerce/request-sender';
-
-import { createCheckoutClient, createCheckoutStore, CheckoutActionCreator, CheckoutRequestSender } from '../checkout';
-import { ConfigActionCreator, ConfigRequestSender } from '../config';
+import { createCheckoutClient, createCheckoutStore } from '../checkout';
 import { PaymentMethodActionCreator } from '../payment';
 
 import CheckoutButtonInitializer from './checkout-button-initializer';
@@ -10,16 +7,11 @@ import createCheckoutButtonRegistry from './create-checkout-button-registry';
 
 export default function createCheckoutButtonInitializer(): CheckoutButtonInitializer {
     const store = createCheckoutStore();
-    const requestSender = createRequestSender();
 
     return new CheckoutButtonInitializer(
         store,
         new CheckoutButtonStrategyActionCreator(
             createCheckoutButtonRegistry(store),
-            new CheckoutActionCreator(
-                new CheckoutRequestSender(requestSender),
-                new ConfigActionCreator(new ConfigRequestSender(requestSender))
-            ),
             new PaymentMethodActionCreator(createCheckoutClient())
         )
     );
