+++ src/checkout-buttons/create-checkout-button-registry.ts
@@ -1,8 +1,10 @@
 import { createFormPoster } from '@bigcommerce/form-poster';
+import { createRequestSender } from '@bigcommerce/request-sender';
 import { getScriptLoader } from '@bigcommerce/script-loader';
 
-import { CheckoutStore } from '../checkout';
+import { CheckoutActionCreator, CheckoutRequestSender, CheckoutStore } from '../checkout';
 import { Registry } from '../common/registry';
+import { ConfigActionCreator, ConfigRequestSender } from '../config';
 import { BraintreeScriptLoader, BraintreeSDKCreator } from '../payment/strategies/braintree';
 import { PaypalScriptLoader } from '../payment/strategies/paypal';
 
@@ -10,10 +12,16 @@ import { BraintreePaypalButtonStrategy, CheckoutButtonStrategy } from './strateg
 
 export default function createCheckoutButtonRegistry(store: CheckoutStore): Registry<CheckoutButtonStrategy> {
     const registry = new Registry<CheckoutButtonStrategy>();
+    const requestSender = createRequestSender();
+    const checkoutActionCreator = new CheckoutActionCreator(
+        new CheckoutRequestSender(requestSender),
+        new ConfigActionCreator(new ConfigRequestSender(requestSender))
+    );
 
     registry.register('braintreepaypal', () =>
         new BraintreePaypalButtonStrategy(
             store,
+            checkoutActionCreator,
             new BraintreeSDKCreator(new BraintreeScriptLoader(getScriptLoader())),
             new PaypalScriptLoader(getScriptLoader()),
             createFormPoster()
@@ -23,6 +31,7 @@ export default function createCheckoutButtonRegistry(store: CheckoutStore): Regi
     registry.register('braintreepaypalcredit', () =>
         new BraintreePaypalButtonStrategy(
             store,
+            checkoutActionCreator,
             new BraintreeSDKCreator(new BraintreeScriptLoader(getScriptLoader())),
             new PaypalScriptLoader(getScriptLoader()),
             createFormPoster(),
